deleted/refactored main/login to main
