+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,10 @@
             android:name=".ForgotPassActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
-            android:exported="false" />
-        <activity
             android:name=".RegisterActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".MainActivity"
             android:exported="false" />
         <activity
             android:name=".HomeActivity"
